+++ MacPacker/CHANGELOG.md
@@ -1,3 +1,6 @@
+v0.11
+- fix: Crash when using the open with option
+
 v0.10
 - feat: Quick Look extension
 - feat: Added Send-a-smile menu to let users quickly star the repo or add a review in the App Store